pyproject.toml
[project]
name = "cpn-telegram-bot"
version = "0.0.0"
description = "Telegram bot for cpn"
readme = "README.md"
authors = [
  { name = "KevinNitroG", email = "kevinnitro@duck.com" },
  { name = "NTGNguyen", email = 'ntgnguyen@duck.com' },
  { name = "WeeCiCi", email = 'wicici310@gmail.com' },
]
maintainers = [
  { name = "KevinNitroG", email = "kevinnitro@duck.com" },
  { name = "NTGNguyen", email = 'ntgnguyen@duck.com' },
  { name = "WeeCiCi", email = 'wicici310@gmail.com' },
]
requires-python = ">=3.13"
dependencies = [
  "beanie>=1.29.0",
  "cpn-core[ocr]==0.5.0",
  "pydantic>=2.10.5",
  "python-dotenv>=1.0.1",
  "python-telegram-bot>=21.10",
]

[project.scripts]
cpn-telegram-bot = "cpn_telegram_bot:__main__.main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cpn_telegram_bot"]

[dependency-groups]
dev = ["pre-commit>=4.0.1", "commitizen>=4.1.0"]
